Add Descriptor_Type_Any and Operand_Type_Any
Descriptor_Type_Any to be used for type inference purposes and

Operand_Type_Any is necessary to allow the new flow where we
pass result location (operand) to the function doing the operation
(like + or -) to avoid unnecessary copying. Since there are cases
where we do not know the size of the result ahead of time and thus
can not create a temp operand either on the stack or in a register,
we need Operand_Type_Any to alias the source instead and take care
of any necessary copying on the outside.

If the plan works out we would get the benefits of inside-out evaluation,
namely copy elision, without the necessity for consequent optimization
pass that would optimize out unnecessary copying.