Switch from a start location to range for source tracking
Previously we would only have start line / column
and a Slice at that position. This was OK for real
tokens but falls apart during rewriting. This PR
switches to tacking byte offsets so they can be
easily merged and then byte offset to line
conversion is only happening when printing errors.