Remove FIXME about difference between ; terminated an unterminated last statement
After thinking about it it looks like allowing user to put the semicolon on the
last line and still use it as the return result of the block is harmless.

If the expected result is void then it does the value is going to be discarded
anyway and if there is an expected result and the type matches probably
that is what user wants anyway.