Use explicit stack segmentation instead of range based heuristic
Before this commit if the stack offset was negative we assume it
is a local variable on the stack, otherwise it could be an argument
function received on the stack or an argument function puts on the
stack for the downstream call.

The whole setup with on the stack function arguments was only working
by accident and only in limited cases as it was not correctly
accounting for the stack size or volatile register pushes.

To solve this Memory_Location values are now explicitly marked with
the relevant area of the stack - local, received argument or
call target argument.