Remove Operand_Size_Any and deal with LEA in the calling code
`LEA` is a weird instruction in that the size of the operands affects
what the instruction *does*, instead of describing the operands.
For the purposes of this compiler we always want it to generate 64-bit
effective address and then store that full address in the target register.
This is why here we are forcing the source memory operand to be 8 bytes
and check that the target register is also 8 bytes in size.