Make overload scoring more precise
Previously it would weight each parameter match
and compare the overall score. It worked but would be really opaque for the user to understand.

The new system has a much simpler logic:

1. Function with less generic params wins
2. Otherwise function with less casts wins
3. Otherwise function with move exact value matches wins
4. Otherwise compile-time fn wins over runtime

Need to see if that results in nice behavior, but all
the existing tests pass the exact logic can be easily adjusted.